requirements.txt
fastapi==0.120.0
uvicorn[standard]==0.38.0
sqlalchemy[asyncio]==2.0.44
asyncpg==0.30.0
pydantic[email]==2.12.3
pydantic-settings==2.11.0
python-dotenv==1.2.1